tài liệu/Flask-Backend/requirements.txt
requests
urllib3
flask
flask-restx
flask-sock
Flask-SQLAlchemy
Flask-APScheduler
flask-cors
PyJWT
pyotp